requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.9.1
python-dotenv==1.0.0
ytmusicapi==1.3.2
yt-dlp>=2025.05.22
openai-whisper==20231117
openai==1.3.7
ffmpeg-python==0.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
aiofiles==23.2.1
slowapi==0.1.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
psutil==7.0.0
